distance meter as int

+++ pogobot.py
@@ -789,14 +789,14 @@ def sendOnePoke(chat_id, pokemon):
             if pref.get('walk_dist'):
                 walkin_data = get_walking_data(location_data, latitude, longitude)
                 if walkin_data['walk_dist'] < 1:
-                    title += " 📍 %sm" % (1000*walkin_data['walk_dist'])
+                    title += " 📍 %sm" % (int(1000*walkin_data['walk_dist']))
                 else:
                     title += " 📍 %skm" % (walkin_data['walk_dist'])
                 address += " 🚶%s" % (walkin_data['walk_time'])
             else:
                 dist = round(pokemon.getDistance(location_data), 2)
                 if dist < 1:
-                    title += " 📍 %sm" % (1000*dist)
+                    title += " 📍 %sm" % (int(1000*dist))
                 else:
                     title += " 📍 %skm" % (dist)
